pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-document-organizer"
version = "1.1.0"
description = "An intelligent document organization application powered by AI (Cloud & Local-First modes)"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Win32 (MS Windows)",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business",
    "Topic :: Text Processing :: General",
]
dependencies = [
    "google-generativeai>=0.3.1",
    "openai>=1.3.0",
    "pandas>=1.5.3",
    "openpyxl>=3.1.2",
    "python-docx>=0.8.11",
    "beautifulsoup4>=4.12.2",
    "markdown>=3.4.3",
    "pillow>=9.5.0",
    "tqdm>=4.66.1",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    # Local-First mode dependencies
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "pdfplumber>=0.10.0",
    "python-pptx>=0.6.21",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]
packaging = [
    "pyinstaller>=5.13.0",
    "cx_freeze>=6.15.0",
]
windows = [
    "pywin32>=306",
]

[project.scripts]
organize = "src.organizer.cli:main"
smart-organizer = "src.organizer.cli:main"

[project.urls]
"Homepage" = "https://github.com/yourusername/ai-document-organizer"
"Bug Tracker" = "https://github.com/yourusername/ai-document-organizer/issues"

[tool.setuptools]
packages = ["src", "src.organizer"]
package-dir = {"" = "."}

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"
markers = [
    "unit: Unit tests (fast, no external dependencies)",
    "integration: Integration tests (may require Ollama)",
    "slow: Slow tests (large files, real LLM calls)",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
